[251] LorrisAnalyzer: export setDataType() from Number and Bar widgets to script

+++ src/LorrisAnalyzer/DataWidgets/GraphWidget/graphdata.cpp
@@ -205,7 +205,7 @@ quint32 GraphData::getStorageBegin(qint32 absPos)
     }
     else if((qint32)m_data.size() < m_sample_size)
     {
-        if(m_data_pos > m_sample_size)
+        if(m_data_pos > (quint32)m_sample_size)
             return m_data_pos - m_sample_size + m_data.size();
 
     }
